added market.vendors and tests

+++ lib/sale.rb
@@ -30,7 +30,7 @@ module FarMar
     
     def self.all
       sales = CSV.readlines("support/sales.csv").map do |line|
-        Sale.new(line[0].to_i, line[1].to_f, Time.parse(line[2]), line[3].to_i, line[4].to_i)
+        Sale.new(line[0].to_i, line[1].to_i, Time.parse(line[2]), line[3].to_i, line[4].to_i)
       end
       return sales
     end
